Firefox browser install now uses mozdownload and mozinstall
mozinstall will refuse to install if firefox is already installed
(when checked on OSX). I've currently only set it to print an error message
when this happens - it'll require the user to uninstall firefox in _venv and run
the command again as uninstall in mozinstall doesn't seem to work very well.

Closes https://github.com/w3c/web-platform-tests/issues/7996